Rename component users to org-users
Currently the "users" deals with users within the context of an
organisation. When we add a component for managing users outside the
context of an organisation we want to call it "users" which, unless we
rename the current component, is a conflict.
