--- src/components/users/index.ts
@@ -1,1 +1,0 @@
-export * from './users';
